+++ flutter_legado/lib/src/screens/book_info_screen.dart
@@ -907,6 +907,7 @@ class _BookInfoScreenState extends ConsumerState<BookInfoScreen> {
             width: 110,
             height: 160,
             borderRadius: 10,
+            sourceOrigin: book.origin,
           ),
         ),
       ),